feat(backend): adjust to new database schema
